+++ httprest/http/base.py
@@ -7,6 +7,7 @@ from typing import Optional
 from urllib.parse import urlencode
 
 from .auth import BaseAuth
+from .cert import ClientCertificate
 
 
 class HTTPRequestError(Exception):
@@ -83,6 +84,7 @@ class HTTPClient(ABC):
         url: str,
         json: Optional[dict] = None,
         headers: Optional[dict] = None,
+        cert: Optional[ClientCertificate] = None,
     ) -> HTTPResponse:
         # pylint: disable=too-many-arguments
         """Perform request.
@@ -91,6 +93,10 @@ class HTTPClient(ABC):
         as `HTTPRequestError`.
 
         Headers may be extended if necessary.
+
+        .. warning:: there is no unified handling for `cert`. The client must
+          raise NotImplementedError if client side certificates are not
+          supported
         """
 
     def request(
@@ -101,6 +107,7 @@ class HTTPClient(ABC):
         headers: Optional[dict] = None,
         params: Optional[dict] = None,
         auth: Optional[BaseAuth] = None,
+        cert: Optional[ClientCertificate] = None,
     ) -> HTTPResponse:
         # pylint: disable=too-many-arguments
         """Perform HTTP request with a given HTTP method.
@@ -112,10 +119,11 @@ class HTTPClient(ABC):
         :param params: query parameters. If provided, the url will be extended
         :param auth: authorization to use. If provided, the headers will be
           extended
+        :param cert: client side certificate
         """
         logging.info("%s %s", method.upper(), url)
         if auth:
             headers = auth.apply(headers or {})
         if params:
             url = f"{url}?{urlencode(params)}"
-        return self._request(method, url, json, headers=headers)
+        return self._request(method, url, json, headers=headers, cert=cert)